ai/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.100.2
pymilvus==2.6.0
pymilvus.model==0.3.2
nltk==3.9.1
numpy==2.0.2
python-dotenv==1.1.1
voyageai==0.3.4
